Split secret store test to isolate boto3 lib requirements
